update assets in example instead of build
